Move OCLCClassifyAPI to appropriate file & fix imports
