+++ oclc.py
@@ -4,7 +4,6 @@ import datetime
 import json
 import logging
 import re
-import urllib
 
 import isbnlib
 from collections import Counter
@@ -675,35 +674,6 @@ class OCLCLinkedData(object):
                     for graph in self.graphs_for(i.output):
                         yield graph
         self.log.debug("END GRAPHS FOR %r", identifier)
-
-
-class OCLCClassifyAPI(object):
-
-    BASE_URL = 'http://classify.oclc.org/classify2/Classify?'
-
-    NO_SUMMARY = '&summary=false'
-
-    def __init__(self, _db):
-        self._db = _db
-
-    @property
-    def source(self):
-        return DataSource.lookup(self._db, DataSource.OCLC)
-
-    def query_string(self, **kwargs):
-        args = dict()
-        for k, v in kwargs.items():
-            if isinstance(v, unicode):
-                v = v.encode("utf8")
-            args[k] = v
-        return urllib.urlencode(sorted(args.items()))
-
-    def lookup_by(self, **kwargs):
-        """Perform an OCLC Classify lookup."""
-        query_string = self.query_string(**kwargs)
-        url = self.BASE_URL + query_string
-        representation, cached = Representation.get(self._db, url)
-        return representation.content
 
 
 class OCLCXMLParser(XMLParser):
